Move entities interfaces from tag bundle into component
