+++ spec/Badger/TagBundle/Security/TaggedEntityVoterSpec.php
@@ -2,7 +2,7 @@
 
 namespace spec\Badger\TagBundle\Security;
 
-use Badger\TagBundle\Entity\TagInterface;
+use Badger\Component\Tag\Model\TagInterface;
 use Badger\TagBundle\Security\TaggedEntityVoter;
 use Badger\TagBundle\Taggable\TaggableInterface;
 use Badger\UserBundle\Entity\User;
