+++ src/Badger/Bundle/GameBundle/Entity/Adventure.php
@@ -5,7 +5,7 @@ namespace Badger\GameBundle\Entity;
 use Badger\Component\Game\Model\AdventureInterface;
 use Badger\Component\Game\Model\AdventureStepInterface;
 use Badger\Component\Game\Model\BadgeInterface;
-use Badger\TagBundle\Entity\TagInterface;
+use Badger\Component\Tag\Model\TagInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 
 /**
