+++ src/Badger/Bundle/GameBundle/Entity/Badge.php
@@ -3,7 +3,7 @@
 namespace Badger\GameBundle\Entity;
 
 use Badger\Component\Game\Model\BadgeInterface;
-use Badger\TagBundle\Entity\TagInterface;
+use Badger\Component\Tag\Model\TagInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use JsonSerializable;
