+++ src/Badger/Component/Game/Repository/TagSearchableRepositoryInterface.php
@@ -2,7 +2,7 @@
 
 namespace Badger\Component\Game\Repository;
 
-use Badger\TagBundle\Entity\TagInterface;
+use Badger\Component\Tag\Model\TagInterface;
 
 /**
  * Repository interface for tag searchable entities.
